requirements.txt
Django
django-countries
django_ckeditor
